pyproject.toml
[tool.poetry]
name = "newsfeed"
version = "0.1.0"
description = "TODO"
authors = ["Modulai <info@modulai.io>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
jsonargparse = "^4.22.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
loguru = "^0.7.0"
pandas = "^2.0.3"
pydantic = "^1"
langchain = "^0.0.221"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
black = "^23.3.0"
isort = "^5.12.0"
pre-commit = "^3.3.1"
mypy = "^1.3.0"
pytest = "^7.4.0"
discord = "^2.3.1"
apache-airflow = "^2.6.2"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
